refactor(core): mime types stream
