+++ lib/httpInterface.js
@@ -21,14 +21,10 @@
 const Ajv = require('ajv')
 const ajvFormats = require('ajv-formats')
 const ajvKeywords = require('ajv-keywords')
-const ndjson = require('ndjson')
-const csvParse = require('csv-parse')
-const csvStringify = require('csv-stringify')
-const fastJson = require('fast-json-stringify')
+
 
 const { pipeline } = require('stream/promises')
 const lget = require('lodash.get')
-const JSONStream = require('JSONStream')
 const through2 = require('through2')
 
 const lisEmpty = require('lodash.isempty')
@@ -56,6 +52,7 @@ const BadRequestError = require('./BadRequestError')
 const { SCHEMAS_ID } = require('./schemaGetters')
 const { getAjvResponseValidationFunction, shouldValidateStream, shouldValidateItem } = require('./validatorGetters')
 const BatchWritableStream = require('./BatchWritableStream')
+const { getFileMimeParser, getFileMimeStringify } = require('./mimeTypeTransform')
 
 const BAD_REQUEST_ERROR_STATUS_CODE = 400
 const UNSUPPORTED_MIME_TYPE_STATUS_CODE = 415
@@ -1187,63 +1184,3 @@ function getEditableFields(aclWriteColumns, allFieldNames) {
   return editableFields.filter(ef => !internalFields.includes(ef))
 }
 
-function getFileMimeStringify(contentType, parsingOptions = {}) {
-  const mimeStringify = {
-    'application/x-ndjson': { stringify: (schema) => {
-      const parse = fastJson(schema)
-      return through2.obj((obj, _encoding, callback) => {
-        callback(null, `${parse(obj)}\n`)
-      })
-    },
-    args: [parsingOptions.schema] },
-    'application/json': { stringify: JSONStream.stringify, args: [] },
-    'text/csv': { stringify: csvStringify.stringify,
-      args: [{
-        encoding: 'utf8',
-        delimiter: ';',
-        escape: '\\',
-        header: true,
-        quote: false,
-        ...parsingOptions,
-        cast: {
-          object: (value) => {
-            try {
-              return { value: JSON.stringify(value), quote: true }
-            } catch (errs) {
-              return value
-            }
-          },
-        },
-      }],
-    },
-  }
-
-  return mimeStringify[contentType]
-}
-
-function getFileMimeParser(contentType, parsingOptions) {
-  const mimeParser = {
-    'application/x-ndjson': { parser: ndjson.parse, args: [] },
-    'application/json': { parser: JSONStream.parse, args: ['*'] },
-    'text/csv': { parser: csvParse.parse,
-      args: [{
-        encoding: 'utf8',
-        delimiter: ';',
-        columns: true,
-        skip_empty_lines: true,
-        relax_quotes: true,
-        escape: '\\',
-        ...parsingOptions,
-        cast: (value) => {
-          try {
-            return JSON.parse(value)
-          } catch (errs) {
-            return value
-          }
-        },
-      }],
-    },
-  }
-
-  return mimeParser[contentType]
-}
